Organize scripts into package
