+++ Makefile
@@ -1,7 +1,7 @@
 PY = $$(which python3 || which python)
 URL = "https://github.com/kerrigan29a/microdoc/blob/main/{path}\#L{line}"
 
-SOURCES = $(wildcard *.py)
+SOURCES = $(wildcard microdoc/*.py)
 
 .PHONY: all clean test
 
@@ -11,10 +11,10 @@ test:
 	$(PY) -m unittest discover -v
 
 README.md: README.md.in $(SOURCES)
-	(cat README.md.in | \
-		$(PY) inject_usage.py) > README.md
-	( $(PY) py2doc.py $(SOURCES) | \
-		$(PY) doc2md.py -l 2 -u $(URL) ) >> README.md
+       (cat README.md.in | \
+               $(PY) -m microdoc.inject_usage) > README.md
+       ( $(PY) -m microdoc.py2doc $(SOURCES) | \
+               $(PY) -m microdoc.doc2md -l 2 -u $(URL) ) >> README.md
 
 clean:
 	rm -f README.md *.json
