+++ microdoc/__init__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+"""Microdoc package exposing the command line tools."""
+
+from ._version import __version__
+
+from . import doc2md
+from . import md2html
+from . import py2doc
+from . import doctest_utils
+
+__all__ = [
+    "__version__",
+    "doc2md",
+    "md2html",
+    "py2doc",
+    "doctest_utils",
+]
