+++ microdoc/doctest_utils.py
@@ -9,7 +9,7 @@ Extension of [doctest](https://docs.python.org/3/library/doctest.html) to allow
 testing Markdown texts.
 """
 
-from _version import __version__
+from ._version import __version__
 
 from doctest import (
     DocTestFinder,
