+++ microdoc/inject_usage.py
@@ -9,15 +9,15 @@ import subprocess
 
 for line in sys.stdin:
     if "{{PY2DOC_USAGE}}" in line:
-        output = subprocess.check_output([sys.executable, "py2doc.py", "-h"])
+        output = subprocess.check_output([sys.executable, "-m", "microdoc.py2doc", "-h"])
         line = line.replace("{{PY2DOC_USAGE}}", output.rstrip().decode("utf-8"))
         line = f"```\n{line}```\n"
     elif "{{DOC2MD_USAGE}}" in line:
-        output = subprocess.check_output([sys.executable, "doc2md.py", "-h"])
+        output = subprocess.check_output([sys.executable, "-m", "microdoc.doc2md", "-h"])
         line = line.replace("{{DOC2MD_USAGE}}", output.rstrip().decode("utf-8"))
         line = f"```\n{line}```\n"
     elif "{{MD2HTML_USAGE}}" in line:
-        output = subprocess.check_output([sys.executable, "md2html.py", "-h"])
+        output = subprocess.check_output([sys.executable, "-m", "microdoc.md2html", "-h"])
         line = line.replace("{{MD2HTML_USAGE}}", output.rstrip().decode("utf-8"))
         line = f"```\n{line}```\n"
     sys.stdout.write(line)
